# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trivia game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List games, optionally by status
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [waiting, active, completed]
      responses:
        "200":
          description: Games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createGame
      tags: [games]
      summary: Create a game (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{gameId}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a game with participant and question counts
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: The game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameDetail"
        "404":
          description: Not found
    patch:
      operationId: updateGame
      tags: [games]
      summary: Update a game (status, topic, difficulty)
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Not found
    delete:
      operationId: deleteGame
      tags: [games]
      summary: Delete a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /games/{gameId}/questions/import-opentdb:
    post:
      operationId: importOpenTdbQuestions
      tags: [questions]
      summary: Import questions from Open Trivia Database (admin)
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenTdbImportInput"
      responses:
        "200":
          description: Questions imported successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenTdbImportResult"
        "404":
          description: Game not found
        "429":
          description: Rate limited by OpenTDB
  /games/{gameId}/questions/generate-gemini:
    post:
      operationId: generateGeminiQuestions
      tags: [questions]
      summary: Generate questions via Google Gemini for custom topics (admin)
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiGenerateInput"
      responses:
        "200":
          description: Questions generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenTdbImportResult"
        "404":
          description: Game not found
        "503":
          description: Gemini API unavailable or key missing
  /games/{gameId}/questions/{questionId}/regenerate:
    post:
      operationId: regenerateQuestion
      tags: [questions]
      summary: Regenerate a single AI question preview using Gemini (not saved)
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
        - name: questionId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateQuestionInput"
      responses:
        "200":
          description: New question preview (not yet saved to DB)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegenerateQuestionPreview"
        "404":
          description: Question or game not found
        "503":
          description: Gemini unavailable or key missing
  /games/{gameId}/questions/{questionId}/enhance:
    post:
      operationId: enhanceQuestion
      tags: [questions]
      summary: Analyze a question and suggest improvements using Gemini
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
        - name: questionId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Enhancement suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnhanceQuestionResult"
        "404":
          description: Not found
        "503":
          description: Gemini unavailable
  /games/{gameId}/questions/{questionId}/fact-check:
    post:
      operationId: factCheckQuestion
      tags: [questions]
      summary: Fact-check a single question using Gemini
      parameters:
        - name: gameId
          in: path
          required: true
          schema: { type: integer }
        - name: questionId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Fact check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FactCheckSingleResult"
        "404":
          description: Not found
        "503":
          description: Gemini unavailable
...
components:
  schemas:
...
    GameInput:
      type: object
      required: [topic, difficulty]
      properties:
        topic: { type: string, minLength: 1 }
        difficulty:
          type: string
          enum: [easy, medium, hard]
        createdByAdmin: { type: boolean }
    GameUpdate:
      type: object
      properties:
        topic: { type: string, minLength: 1 }
        difficulty:
          type: string
          enum: [easy, medium, hard]
        status:
          type: string
          enum: [waiting, active, completed]
    Game:
      type: object
      required: [id, topic, difficulty, questionCount, status, createdAt, createdByAdmin]
      properties:
        id: { type: integer }
        topic: { type: string }
        difficulty:
          type: string
          enum: [easy, medium, hard]
        questionCount: { type: integer }
        status:
          type: string
          enum: [waiting, active, completed]
        createdAt: { type: string }
        createdByAdmin: { type: boolean }
    GameDetail:
      type: object
      required: [id, topic, difficulty, questionCount, status, createdAt, createdByAdmin,participantCount]
      properties:
        id: { type: integer }
        topic: { type: string }
        difficulty:
          type: string
          enum: [easy, medium, hard]
        questionCount: { type: integer }
        status:
          type: string
          enum: [waiting, active, completed]
        createdAt: { type: string }
        createdByAdmin: { type: boolean }
        participantCount: { type: integer }
...
    RegenerateQuestionInput:
      type: object
      properties:
        difficulty:
          type: string
          enum: [easy, medium, hard]
        questionType:
          type: string
          enum: [multiple_choice, true_false, write_in]
    RegenerateQuestionPreview:
      type: object
      required: [questionType, questionText, correctAnswer, points, source]
      properties:
        questionType: { type: string }
        questionText: { type: string }
        correctAnswer: { type: string }
        options:
          type: array
          nullable: true
          items: { type: string }
        points: { type: integer }
        source: { type: string }
    EnhanceQuestionResult:
      type: object
      required: [improvedQuestionText, factCheckResult, factCheckNotes, suggestedSource,suggestions]
      properties:
        improvedQuestionText: { type: string }
        improvedOptions:
          type: array
          nullable: true
          items: { type: string }
        factCheckResult:
          type: string
          enum: [VERIFIED, UNCERTAIN, LIKELY_INCORRECT]
        factCheckNotes: { type: string }
        suggestedSource: { type: string }
        suggestions: { type: string }
    FactCheckSingleResult:
      type: object
      required: [verdict, confidence, explanation]
      properties:
        verdict:
          type: string
          enum: [CORRECT, INCORRECT, UNCERTAIN]
        confidence:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        explanation: { type: string }
        correctAnswerIfWrong:
          type: string
          nullable: true
    GeminiGenerateInput:
      type: object
      required: [topic, difficulty, amount]
      properties:
        topic: { type: string, minLength: 1 }
        difficulty:
          type: string
          enum: [easy, medium, hard]
        amount:
          type: integer
          minimum: 1
          maximum: 20
        existingQuestions:
          type: array
          items: { type: string }
    OpenTdbImportInput:
      type: object
      required: [categoryId, difficulty, amount]
      properties:
        categoryId: { type: integer }
        difficulty:
          type: string
          enum: [easy, medium, hard]
        amount:
          type: integer
          minimum: 1
          maximum: 50
    OpenTdbImportResult:
      type: object
      required: [imported, total]
      properties:
        imported: { type: integer }
        total: { type: integer }
...